Renamed pandas_accessor to accessor.utilities
